Implement usage of `enum` as pattern or part of pattern in `match` cases.
