+++ tree.py
@@ -148,7 +148,7 @@ expCaseList:list[ExpCase] = [
 
 patternMatchCasesSolid = [
     MT_Other(),
-    MTVal(), MTString(), MTRegexp(), MTList(), MTTuple(), MTDict(), MTStruct(), 
+    MTVal(), MTString(), MTRegexp(), MTList(), MTTuple(), MTDict(), MTStruct(), MTObjMember(),
 ]
 
 patternMatchCasesCplx = [
@@ -176,6 +176,7 @@ def mtCases(elems:list[Elem], cases: list[ExpCase], parent:ExpCase=None)->Matchi
             if mtCase.hasSubExpr():
                 pattr, subElems = mtCase.split(elems)
                 subExp = elems2expr(subElems)
+                # print('$3>', subExp)
                 pattr = mtCase.setSub(pattr, subExp)
             else:
                 pattr = mtCase.expr(elems)
